+++ Packages/webxr/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- An option in WebXRCameraSettings to set flat/normal mode camera pose when returning from XR mode.
+
 ### Changed
 - XR Hand pinch is now detected manually instead of relying on system select event, to support Apple Vision Pro.